Move translator methods to new internal_gains and model resource files.
