+++ HPXMLtoOpenStudio/measure.xml
@@ -3,8 +3,8 @@
   <schema_version>3.1</schema_version>
   <name>hpxm_lto_openstudio</name>
   <uid>b1543b30-9465-45ff-ba04-1d1f85e763bc</uid>
-  <version_id>91250812-e128-411f-b6d9-1823d155458e</version_id>
-  <version_modified>2024-09-03T22:18:26Z</version_modified>
+  <version_id>1cae9e5f-f16d-42a1-9c30-492d9daa9dcd</version_id>
+  <version_modified>2024-09-03T23:54:58Z</version_modified>
   <xml_checksum>D8922A73</xml_checksum>
   <class_name>HPXMLtoOpenStudio</class_name>
   <display_name>HPXML to OpenStudio Translator</display_name>
@@ -183,7 +183,7 @@
       <filename>measure.rb</filename>
       <filetype>rb</filetype>
       <usage_type>script</usage_type>
-      <checksum>E584F2E5</checksum>
+      <checksum>25D005AC</checksum>
     </file>
     <file>
       <filename>airflow.rb</filename>
@@ -345,13 +345,13 @@
       <filename>geometry.rb</filename>
       <filetype>rb</filetype>
       <usage_type>resource</usage_type>
-      <checksum>08E9B56C</checksum>
+      <checksum>3CF58696</checksum>
     </file>
     <file>
       <filename>hotwater_appliances.rb</filename>
       <filetype>rb</filetype>
       <usage_type>resource</usage_type>
-      <checksum>36092007</checksum>
+      <checksum>B8578044</checksum>
     </file>
     <file>
       <filename>hpxml.rb</filename>
@@ -393,13 +393,19 @@
       <filename>hvac.rb</filename>
       <filetype>rb</filetype>
       <usage_type>resource</usage_type>
-      <checksum>7EC4C04D</checksum>
+      <checksum>93926A7F</checksum>
     </file>
     <file>
       <filename>hvac_sizing.rb</filename>
       <filetype>rb</filetype>
       <usage_type>resource</usage_type>
       <checksum>9B75637E</checksum>
+    </file>
+    <file>
+      <filename>internal_gains.rb</filename>
+      <filetype>rb</filetype>
+      <usage_type>resource</usage_type>
+      <checksum>808BF9F9</checksum>
     </file>
     <file>
       <filename>lighting.rb</filename>
@@ -442,6 +448,12 @@
       <filetype>rb</filetype>
       <usage_type>resource</usage_type>
       <checksum>D04D654F</checksum>
+    </file>
+    <file>
+      <filename>model.rb</filename>
+      <filetype>rb</filetype>
+      <usage_type>resource</usage_type>
+      <checksum>040BA07A</checksum>
     </file>
     <file>
       <filename>output.rb</filename>
